pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raganything"
version = "0.1.0"
description = "All-in-One RAG Framework for Multimodal Document Processing"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "RAG-Anything Contributors"}
]
keywords = ["rag", "retrieval-augmented-generation", "multimodal", "document-processing", "nlp", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
image = [
    "Pillow>=10.0.0",
]
text = [
    "ReportLab>=4.0.0",
]
all = [
    "Pillow>=10.0.0",
    "ReportLab>=4.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pylint>=3.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
ui = [
    "streamlit>=1.28.0",
]

[project.urls]
Homepage = "https://github.com/HKUDS/RAG-Anything"
Documentation = "https://github.com/HKUDS/RAG-Anything#readme"
Repository = "https://github.com/HKUDS/RAG-Anything"

[tool.setuptools.packages.find]
where = ["."]
include = ["raganything*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pylint.messages_control]
disable = ["C0111", "C0114", "C0115", "C0116"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

